Updates docs and usage to limit meridian specification
You can still use A and P if it ends up working for you but
the docs no longer promise it. (You can even try using 17 if you
like because it works for me.)

Also updated "sao paolo" to "sao paolo brazil" for locations
that didn't automatically match that to the right place.
Again, try to be as lax as possible and see what it takes for
it to work for you.